requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-cors-headers
